baybars/requirements.txt
python-consul==0.7.2
azure-storage-blob==0.37.0
azure-storage-queue==0.37.0
confluent-kafka==0.11.0
pydocumentdb==2.3.1
pysftp==0.2.9
requests==2.17.3
numpy==1.14.1
pandas==0.20.3
pymssql==2.1.3